requirements.txt
aiosmtplib==3.0.2
amqp==5.2.0
anyio==4.6.0
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.2.1
boto3==1.26.64
botocore==1.29.84
cachetools==5.3.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==43.0.1
Deprecated==1.2.13
dj-database-url==1.2.0
Django==4.2.7
django-celery-beat==2.7.0
django-cors-headers==3.13.0
django-countries-plus==2.1.0
django-dotenv==1.4.2
django-filter==22.1
django-storages==1.13.2
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
exceptiongroup==1.2.2
filelock==3.16.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.11.2
google-cloud-vision==3.4.4
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.26.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.4.2
Markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
paho-mqtt==2.1.0
phonenumbers==8.12.48
Pillow==9.4.0
pkce==1.0.3
prompt_toolkit==3.0.48
proto-plus==1.22.3
protobuf==4.23.4
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.10.4
pygame==2.6.1
PyJWT==2.6.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
redis==5.1.0
requests==2.28.2
requests-file==2.1.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.3
tldextract==3.1.0
twilio==7.9.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.3.0
wrapt==1.16.0
